refactor: modularize utilities and update validator
